Lift filter controls out of `SpansView`
This way the parent component can decide which filters are available.
