+++ static/app/views/starfish/views/spans/index.tsx
@@ -1,11 +1,17 @@
+import styled from '@emotion/styled';
+
 import * as Layout from 'sentry/components/layouts/thirds';
+import PageFilterBar from 'sentry/components/organizations/pageFilterBar';
 import {t} from 'sentry/locale';
+import {space} from 'sentry/styles/space';
 import {
   PageErrorAlert,
   PageErrorProvider,
 } from 'sentry/utils/performance/contexts/pageError';
 import {useLocation} from 'sentry/utils/useLocation';
+import StarfishDatePicker from 'sentry/views/starfish/components/datePicker';
 import {StarfishPageFiltersContainer} from 'sentry/views/starfish/components/starfishPageFiltersContainer';
+import {StarfishProjectSelector} from 'sentry/views/starfish/components/starfishProjectSelector';
 import {ModuleName, SpanMetricsFields} from 'sentry/views/starfish/types';
 import {ROUTE_NAMES} from 'sentry/views/starfish/utils/routeNames';
 
@@ -42,6 +48,11 @@ export default function Spans() {
           <Layout.Main fullWidth>
             <PageErrorAlert />
             <StarfishPageFiltersContainer>
+              <StyledPageFilterBar condensed>
+                <StarfishProjectSelector />
+                <StarfishDatePicker />
+              </StyledPageFilterBar>
+
               <SpansView moduleName={moduleName} spanCategory={spanCategory} />
             </StarfishPageFiltersContainer>
           </Layout.Main>
@@ -75,3 +86,8 @@ const getTitle = (moduleName: ModuleName, spanCategory?: string) => {
   }
   return t('Spans');
 };
+
+const StyledPageFilterBar = styled(PageFilterBar)`
+  margin: 0 ${space(2)};
+  margin-bottom: ${space(2)};
+`;
