+++ static/app/views/starfish/views/spans/spansView.tsx
@@ -2,13 +2,10 @@ import {Fragment} from 'react';
 import styled from '@emotion/styled';
 import pick from 'lodash/pick';
 
-import PageFilterBar from 'sentry/components/organizations/pageFilterBar';
 import {space} from 'sentry/styles/space';
 import {fromSorts} from 'sentry/utils/discover/eventView';
 import type {Sort} from 'sentry/utils/discover/fields';
 import {useLocation} from 'sentry/utils/useLocation';
-import StarfishDatePicker from 'sentry/views/starfish/components/datePicker';
-import {StarfishProjectSelector} from 'sentry/views/starfish/components/starfishProjectSelector';
 import {ModuleName, SpanMetricsFields} from 'sentry/views/starfish/types';
 import {QueryParameterNames} from 'sentry/views/starfish/views/queryParameters';
 import {ActionSelector} from 'sentry/views/starfish/views/spans/selectors/actionSelector';
@@ -56,11 +53,6 @@ export default function SpansView(props: Props) {
 
   return (
     <Fragment>
-      <StyledPageFilterBar condensed>
-        <StarfishProjectSelector />
-        <StarfishDatePicker />
-      </StyledPageFilterBar>
-
       <PaddedContainer>
         <SpanTimeCharts
           moduleName={moduleName}
@@ -115,9 +107,4 @@ const FilterOptionsContainer = styled(PaddedContainer)`
   gap: ${space(1)};
   margin-bottom: ${space(2)};
   max-width: 800px;
-`;
-
-const StyledPageFilterBar = styled(PageFilterBar)`
-  margin: 0 ${space(2)};
-  margin-bottom: ${space(2)};
 `;
